Refactor a bit for ease of editing
